engine/requirements.txt
torch
torchvision
pyyaml
numpy
matplotlib
tqdm
tensorboard
opencv-python-headless
timm
torchinfo
seaborn
pandas

pip install torch torchvision pyyaml numpy matplotlib tqdm tensorboard opencv-python-headless timm torchinfo seaborn pandas nibabel gdown albumentations scikit-learn scikit-image monai segmentation_models_pytorch torchmetrics 